refactor : (숫자 야구) 볼, 스트라이크 구하는 로직 분리
