+++ src/main/java/baseball/BaseballGame.java
@@ -1,24 +1,25 @@
 package baseball;
 
 import baseball.constants.SystemConsole;
-import baseball.constants.SystemMessage;
+import baseball.domain.GameResultScore;
 import baseball.domain.RandomGenerator;
 import baseball.domain.Referee;
 import baseball.player.Player;
-import camp.nextstep.edu.missionutils.Console;
 
 import java.util.List;
 
+import static baseball.domain.GameResultScore.getRestart;
+
 public class BaseballGame {
     private static final String RESTART = "1";
-    private static final String PATTERN = "[1-2]+";
     private static final String GAME_WIN_CONDITIONS = "3스트라이크";
-    private static final String RESTART_INPUT = "플레이어가 숫자 1 또는 2가 아닌 다른 값을 입력하였습니다.";
 
     private final SystemConsole systemConsole;
+    private final GameResultScore gameResultScore;
 
-    public BaseballGame(SystemConsole systemConsole) {
+    public BaseballGame(SystemConsole systemConsole, GameResultScore gameResultScore) {
         this.systemConsole = systemConsole;
+        this.gameResultScore = gameResultScore;
     }
 
     public void start() {
@@ -30,7 +31,7 @@ public class BaseballGame {
         conditionsBaseballGame(player, referee, randomGenerator);
     }
 
-    private void conditionsBaseballGame(Player player, Referee referee, RandomGenerator randomGenerator) {
+    public void conditionsBaseballGame(Player player, Referee referee, RandomGenerator randomGenerator) {
         String restartNumber = "1";
 
         while (restartNumber.equals(RESTART)) {
@@ -40,39 +41,13 @@ public class BaseballGame {
         }
     }
 
-    private void guessComputerNumbers(Player player, Referee referee, List<Integer> computer) {
+    public void guessComputerNumbers(Player player, Referee referee, List<Integer> computer) {
         String gameResultNumber = "";
 
         while (!(gameResultNumber.equals(GAME_WIN_CONDITIONS))) {
             systemConsole.PlayerInputMessage();
-            gameResultNumber = getGameResultNumber(player, referee, computer);
+            gameResultNumber = gameResultScore.getGameNumber(player, referee, computer);
             systemConsole.gameResultNumberMessage(gameResultNumber);
         }
-    }
-
-    private String getGameResultNumber(Player player, Referee referee, List<Integer> computer) {
-        String gameResultNumber;
-        List<Integer> playerInput = getPlayerInput(player);
-        gameResultNumber = referee.resultJudgment(computer, playerInput);
-        return gameResultNumber;
-    }
-
-    private List<Integer> getPlayerInput(Player player) {
-        String playerNumber = Console.readLine();
-        return player.getInput(playerNumber);
-    }
-
-    private static String playerRestartInput() {
-        String restart = Console.readLine();
-        if (!(restart.matches(PATTERN))) {
-            throw new IllegalArgumentException(RESTART_INPUT);
-        }
-        return restart;
-    }
-
-    public static String getRestart() {
-        SystemConsole.gameWinMessage();
-        SystemConsole.gameRestartMessage();
-        return playerRestartInput();
     }
 }
